Restructure folder towards proposed standard
