+++ packages/harpstrata/src/harpstrata/IsActive/utils/utils/activeIdsFromMatrixReducers/index.test.ts
@@ -1,7 +1,7 @@
-import { PitchIds, getPitch } from '../../../Pitch'
-import type { PitchMatrix } from '../../../Pitch'
-import { DegreeIds, getDegree } from '../../../Degree'
-import type { DegreeMatrix } from '../../../Degree'
+import { PitchIds, getPitch } from '../../../../Pitch'
+import type { PitchMatrix } from '../../../../Pitch'
+import { DegreeIds, getDegree } from '../../../../Degree'
+import type { DegreeMatrix } from '../../../../Degree'
 
 import type { MatrixAccumulator } from './index'
 import { activeIdsFromPitchMatrix, activeIdsFromDegreeMatrix } from './index'
