+++ packages/harpstrata/src/harpstrata/IsActive/utils/utils/activeIdsFromMatrixReducers/index.ts
@@ -3,9 +3,9 @@ import {
   activeIdsFromPitchRow,
   activeIdsFromDegreeRow,
 } from '../activeIdsFromRowReducers'
-import type { ActivePitchIds, ActiveDegreeIds } from '../../types'
-import type { PitchRow, PitchMatrix } from '../../../Pitch'
-import type { DegreeRow, DegreeMatrix } from '../../../Degree'
+import type { ActivePitchIds, ActiveDegreeIds } from '../../../types'
+import type { PitchRow, PitchMatrix } from '../../../../Pitch'
+import type { DegreeRow, DegreeMatrix } from '../../../../Degree'
 
 export type MatrixAccumulator = {
   readonly degreeMatrix: DegreeMatrix
