+++ packages/harpstrata/src/harpstrata/IsActive/utils/utils/getCounterpartActiveIds/index.test.ts
@@ -1,6 +1,6 @@
-import type { IsActiveProps } from '../../types'
-import { PitchIds, getPitch } from '../../../Pitch'
-import { DegreeIds, getDegree } from '../../../Degree'
+import type { IsActiveProps } from '../../../types'
+import { PitchIds, getPitch } from '../../../../Pitch'
+import { DegreeIds, getDegree } from '../../../../Degree'
 
 import { getCounterpartDegreeIds, getCounterpartPitchIds } from './index'
