+++ packages/harpstrata/src/harpstrata/IsActive/utils/utils/index.ts
@@ -1,0 +1,4 @@
+export {
+  getCounterpartDegreeIds,
+  getCounterpartPitchIds,
+} from './getCounterpartActiveIds'
